ASRC-27: insert_base_data.sql is now only a test data set
Previously, we were configuring the VA FY2013 models in insert_base_data.sql.
But now we have decided to make that file just test data for integration tests
and the FY2013 models will be considered VA-sensitive data until we are
otherwise informed.

Thus the actual variable definition for ASRC-27 will not be pushed to GitHub for
now and this commit just removes the expectation that it would be.