ASRC-199: refactor global SrcalcInfo object
Previously, we used a Spring WebRequestInterceptor to add the SrcalcInfo object
to the Model immediately after the Controller handled each request. Not only was
this a little inefficient, but it meant that only Spring views could access the
object.

With this change, the SrcalcInfo object is added to the ServletContext once,
after which it is accessible from any JSP via the applicationScope.