ASRC-230: improve AbstractNumericalVariable interface
Previously, AbstractNumericalVariable presented various getters and setters for
modifying compnents of the valid range. Internally it modified the
NumericalRange member, which is supposed to be immutable.

This commit tightens up the AbstractNumericalVariable interface by providing a
getValidRange() to access the stored NumericalRange and a setValidRange() to
completely replace it with another immutable instance.