ASRC-52: save results upon signature
Calculation metadata is captured in the HistoricalCalculation class and the actual results
in the SignedResult class. Both are persisted upon signature. For ASRC-59, we will have to
persist HistoricalCalculation immediately after the first run, but I encountered a
Hibernate issue when trying to do that here so that will have to wait.